Move the eff package to the root folder of the repo, updated the examples to import this
