+++ examples/drawing-primitives/circled_dots.go
@@ -4,7 +4,7 @@ import (
 	"math"
 	"math/rand"
 
-	"github.com/forestgiant/eff/eff"
+	"github.com/forestgiant/eff"
 )
 
 type twoPositionDot struct {
