+++ examples/drawing-primitives/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"strconv"
 
-	"github.com/forestgiant/eff/eff"
+	"github.com/forestgiant/eff"
 	"github.com/forestgiant/eff/sdl"
 )
