+++ examples/drawing-primitives/rects.go
@@ -3,7 +3,7 @@ package main
 import (
 	"math/rand"
 
-	"github.com/forestgiant/eff/eff"
+	"github.com/forestgiant/eff"
 )
 
 type rects struct {
