+++ examples/moving-image/main.go
@@ -6,7 +6,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/forestgiant/eff/eff"
+	"github.com/forestgiant/eff"
 	"github.com/forestgiant/eff/sdl"
 )
